requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pandas
openpyxl
google-genai
python-dotenv
